includ last design advice from brandon
