+++ src/assets/svg/SvgIndex.jsx
@@ -47,6 +47,21 @@ export const AlgoliaLogoMobile = memo(() => (
   </svg>
 ))
 
+export const InfoPicto = memo(() => (
+  <svg
+    width="26"
+    height="26"
+    viewBox="0 0 26 26"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M13.2983 25.3747C6.22344 25.3465 0.601203 19.5529 0.806215 12.4951C1.00204 5.75211 6.52303 0.31793 13.3978 0.36604C20.3211 0.414303 25.869 6.07138 25.8001 12.9718C25.7316 19.8514 20.0932 25.4453 13.2977 25.3753L13.2983 25.3747ZM13.5635 11.0901C12.2963 11.1317 11.1634 11.5703 10.1172 12.2621C10.0288 12.3206 9.94397 12.3928 9.87692 12.4746C9.63817 12.7661 9.71996 13.0127 10.0798 13.1067C10.2182 13.1427 10.3602 13.1675 10.4962 13.2111C11.0711 13.3945 11.2262 13.6764 11.0803 14.2724C10.7258 15.7214 10.3643 17.1688 10.0157 18.6196C9.75733 19.6961 10.3025 20.5761 11.3359 20.656C12.8652 20.7739 14.23 20.2693 15.4577 19.3835C15.5878 19.2897 15.6886 19.0565 15.6747 18.8991C15.6664 18.8037 15.4144 18.7249 15.2668 18.6484C15.2047 18.6162 15.1293 18.6102 15.0601 18.5914C14.3761 18.4086 14.1982 18.1226 14.3639 17.4403C14.7069 16.0253 15.0714 14.6154 15.3918 13.1953C15.4678 12.8592 15.4729 12.4729 15.3877 12.1412C15.197 11.4028 14.5508 11.0613 13.5636 11.0901L13.5635 11.0901ZM16.745 6.75601C16.7469 5.44989 15.7374 4.41953 14.4476 4.41095C13.136 4.40198 12.0829 5.44475 12.0938 6.74081C12.1049 8.03238 13.1313 9.04634 14.4329 9.05083C15.7018 9.05467 16.7429 8.0215 16.745 6.75595V6.75601Z"
+      fill="black"
+    />
+  </svg>
+))
+
 export const Glass = memo(() => (
   <svg
     className="glass"
