+++ src/components/demoGuide/SCSS/demoGuide.scss
@@ -38,7 +38,7 @@
 
 .optionDots__wrapper {
   position: absolute;
-  top: 0.5rem;
+  top: 0rem;
   right: 1rem;
   width: 3rem;
   height: 3rem;
@@ -48,6 +48,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  gap: 0.2rem;
 
   &-active {
     transform: rotate3d(0, 0, 1, -90deg);
@@ -60,6 +61,7 @@
   }
 
   p {
+    @include p-text($white-color);
     font-size: 0.6rem;
     writing-mode: vertical-rl;
     text-orientation: mixed;
@@ -68,7 +70,7 @@
   svg {
     width: 6px;
     path {
-      fill: $secondary-color;
+      fill: $white-color;
     }
   }
 }
