+++ src/components/header/SCSS/header.scss
@@ -6,18 +6,52 @@
 .header {
   text-align: center;
   box-shadow: rgba(0, 0, 0, 0.04) 0px 3px 5px;
-  padding: 2rem 0;
+  padding: 0 0 2rem 0;
   background-color: $background-color;
   .container {
-    width: 90%;
     margin: 0 auto;
-    .container__header-top {
+    .container__header-top{
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      width: 100vw;
+      padding: 0.6rem 3rem;
+      background-color: $tertiary-color;
+      .title{
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        h3{
+          @include titleH3($white-color);
+          text-transform: capitalize;
+        }
+      }
+      .selector-list {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        gap: 2rem;
+        margin-right: 2rem;
+        button{
+          svg{
+            path{
+              fill: $white-color;
+            }
+          }
+          p{
+            @include p-text($white-color);
+            font-size: 0.8rem;
+           
+          }
+        }
+      }
+    }
+    .container__header-mid {
       position: relative;
       align-items: center;
       padding: 0 2rem;
       display: flex;
       justify-content: space-between;
-
       a {
         position: relative;
         width: 100%;
@@ -36,17 +70,35 @@
           height: 100%;
         }
       }
-      &__title {
+      .picto-cart{
+        position: relative;
+        width: 10rem;
         display: flex;
-        justify-content: flex-start;
-        align-items: center;
-        min-height: 5rem;
-        min-width: 40%;
-        h1 {
-          @include titleH3($secondary-color);
-          // font-size: 0.9rem;
-          text-transform: capitalize;
-          margin: 0px;
+        justify-content: flex-end;
+        cursor: pointer;
+        svg{
+          width: 32px;
+        }
+        &__active {
+          animation: bounce 0.3s ease-in-out forwards;
+        }
+
+        .notification-cart{
+          position: absolute;
+          top: -0.5rem;
+          right: -0.5rem;
+          width: 20px;
+          height: 20px;
+          border-radius: 50%;
+          background-color: $secondary-color;
+          display: flex;
+          align-items: center;
+          justify-content: center;
+
+          span{
+            @include p-text($white-color);
+            font-size: 0.8rem;
+          }
         }
       }
     }
@@ -60,6 +112,12 @@
         gap: clamp(0.5rem, -1.12rem + 4.13vw, 2rem);
         li {
           z-index: 50;
+          &:last-child{
+            p{
+
+              color: $secondary-color;
+            }
+          }
           cursor: pointer;
           p {
             position: relative;
@@ -72,60 +130,6 @@
             @include selectedNavItem($secondary-color);
           }
         }
-        .picto-cart {
-          position: relative;
-          .notification-cart {
-            position: absolute;
-            background-color: $secondary-color;
-            top: -1rem;
-            right: -0.6rem;
-            border-radius: 50%;
-            width: 1.7rem;
-            height: 1.7rem;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            transform: scale(1);
-            span {
-              color: $white-color;
-            }
-          }
-          svg {
-            width: 1.9rem;
-            height: 1.9rem;
-          }
-
-          body {
-            font: 1em sans-serif;
-            text-align: center;
-          }
-          .stage {
-            border-bottom: 3px solid #444;
-            display: flex;
-            height: 330px;
-            width: 100%;
-          }
-          .box {
-            align-self: flex-end;
-            animation-duration: 2s;
-            animation-iteration-count: infinite;
-            background-color: #f44336;
-            height: 200px;
-            margin: 0 auto 0 auto;
-            transform-origin: bottom;
-            width: 200px;
-            transition: transform 0.3s ease-in-out;
-          }
-          &__active {
-            animation: bounce 0.3s ease-in-out forwards;
-          }
-        }
-      }
-      .container__header-nav-selectors {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        gap: clamp(0.5rem, -1.12rem + 4.13vw, 2rem);
       }
     }
   }
@@ -134,7 +138,7 @@
 // MOBILE & TABLET
 .header {
   .container-mobile {
-    .container__header-top {
+    .container__header-mid {
       margin-top: 1rem;
       padding: 0 1rem;
 
@@ -145,7 +149,7 @@
         }
       }
     }
-    .container__header-top__logo {
+    .container__header-mid__logo {
       // width: 35% !important;
       svg{
         width: 3rem;
@@ -156,7 +160,7 @@
       justify-content: center;
       align-items: center;
       width: 100%;
-      margin: 0 auto;
+      margin: 0 1rem;
 
       .ais-VoiceSearch {
         right: 8%;
@@ -200,14 +204,6 @@
             font-size: 0.8rem;
           }
         }
-        .container__header-nav-selectors {
-          display: flex;
-          flex-direction: column;
-          align-items: flex-start;
-          justify-content: flex-start;
-          gap: 1.5rem;
-        }
-
         .selected {
           &::before {
             content: '';
