+++ src/components/header/components/HeaderMobile.jsx
@@ -4,7 +4,7 @@ import { useRef, useState, useEffect } from 'react'
 import { Link } from 'react-router-dom'
 
 // Recoil Header State
-import { useRecoilValue, useSetRecoilState } from 'recoil'
+import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil'
 
 import { motion } from 'framer-motion'
 
@@ -30,11 +30,19 @@ import CustomVoiceSearchComponent from '@/components/voicesearch/VoiceSearch'
 import Navigation from './Navigation'
 
 // Custom hook to prevent body from scrolling
-import { clickHamburger } from '@/config/cartFunctions'
-import { shouldHaveVoiceSearch } from '@/config/featuresConfig'
+import {
+  cartClick,
+  cartOpen,
+  cartState,
+  clickHamburger,
+} from '@/config/cartFunctions'
+import {
+  shouldHaveCartFunctionality,
+  shouldHaveVoiceSearch,
+} from '@/config/featuresConfig'
 import usePreventScrolling from '@/hooks/usePreventScrolling'
 import useOutsideClick from '@/hooks/useOutsideClick'
-import { AlgoliaLogoMobile } from '@/assets/svg/SvgIndex'
+import { AlgoliaLogoMobile, CartPicto } from '@/assets/svg/SvgIndex'
 
 const HeaderMobile = ({ mobile, tablet }) => {
   // Import configuration from Recoil
@@ -51,12 +59,26 @@ const HeaderMobile = ({ mobile, tablet }) => {
 
   const rulesApplied = useSetRecoilState(rulesAtom)
 
+  // CART
+  const shouldShowCartIcon = useRecoilValue(shouldHaveCartFunctionality)
+  const [cartOpenValue, setCartOpenValue] = useRecoilState(cartOpen)
+  const cartIcon = useSetRecoilState(cartClick)
+  const [showCart, setShowCart] = useRecoilState(cartState)
+
+  const sumAllArticles = (cart) => {
+    let x = 0
+    cart.map((i, index) => {
+      x += i.qty
+    })
+    return x
+  }
+
   // Prevent body from scrolling when panel is open
   usePreventScrolling(isMenuOpen)
 
   return (
     <div className="container container-mobile">
-      <div className="container__header-top">
+      <div className="container__header-mid">
         {/* Hamburger button to open or close the menu dropdown */}
         <div
           className={`${
@@ -71,7 +93,7 @@ const HeaderMobile = ({ mobile, tablet }) => {
           <span className="hamburger__line"></span>
         </div>
         {/* Logo, which returns to the homepage on click */}
-        <div className="container__header-top__logo">
+        <div className="container__header-mid__logo">
           <Link
             to="/"
             aria-label="Back to homepage"
@@ -86,9 +108,28 @@ const HeaderMobile = ({ mobile, tablet }) => {
           </Link>
         </div>
         {/* For a search box Simple center */}
-        <div className="container__header-top__title">
-          <h1>Demo BoilerPlate</h1>
-        </div>
+        {shouldShowCartIcon && (
+          <div
+            className={
+              cartOpenValue ? 'picto-cart picto-cart__active' : 'picto-cart'
+            }
+            onClick={() => {
+              setCartOpenValue(!cartOpenValue)
+              {
+                mobile && setIsMenuOpen(false)
+              }
+            }}
+            ref={cartIcon}
+          >
+            <CartPicto />
+            {/* Picto notification up the cart icon */}
+            {showCart?.length !== 0 && (
+              <div className="notification-cart">
+                <span>{sumAllArticles(showCart)}</span>
+              </div>
+            )}
+          </div>
+        )}
       </div>
 
       <div className="searchbox-container searchbox-container-mobile">
