+++ src/components/header/components/Navigation.jsx
@@ -13,35 +13,12 @@ import {
   selectorNavigationRef,
 } from '@/config/navigationConfig'
 
-// Import Recoil config
-import {
-  shouldHaveCartFunctionality,
-  shouldHaveLanguages,
-  shouldHavePersona,
-  shouldHaveSegments,
-} from '@/config/featuresConfig'
-
-import { cartOpen, cartState } from '@/config/cartFunctions'
-import { Selectors } from '../../selector/Selectors'
-
-// Import segment configuration
-import { CartPicto } from '@/assets/svg/SvgIndex'
-import { languagesConfig } from '@/config/languagesConfig'
-import { personaConfig } from '@/config/personaConfig'
-import { segmentConfig } from '@/config/segmentConfig'
-import useStoreCartToLocalStorage from '@/hooks/useStoreCartToLocalStorage'
-import { useEffect } from 'react'
-
 //Import config from helped navigation
 import { cartClick } from '@/config/cartFunctions'
 import { windowSize } from '@/hooks/useScreenSize'
 
 const Navigation = ({ isMenuOpen, setIsMenuOpen }) => {
-  // Recoil State
-  const [cartOpenValue, setCartOpenValue] = useRecoilState(cartOpen)
-  const [showCart, setShowCart] = useRecoilState(cartState)
-
-  const { mobile, isDesktop } = useRecoilValue(windowSize)
+  const { isDesktop } = useRecoilValue(windowSize)
 
   // navigate is used by React Router
   const navigate = useNavigate()
@@ -66,15 +43,6 @@ const Navigation = ({ isMenuOpen, setIsMenuOpen }) => {
     }
   }
 
-  // Get references for dropdowns in Navigation
-  const selectorsNavigation = useSetRecoilState(selectorNavigationRef)
-
-  // Should show or not the sections
-  const shouldShowPersonasAtom = useRecoilValue(shouldHavePersona)
-  const shouldShowSegmentsAtom = useRecoilValue(shouldHaveSegments)
-  const shouldShowLanguageSelected = useRecoilValue(shouldHaveLanguages)
-  const shouldShowCartIcon = useRecoilValue(shouldHaveCartFunctionality)
-
   // Import the navigation links, as defined in the config
   const links = useRecoilValue(linksHeader)
 
@@ -82,34 +50,6 @@ const Navigation = ({ isMenuOpen, setIsMenuOpen }) => {
 
   const [navigationState, setNavigationState] =
     useRecoilState(navigationStateAtom)
-
-  const cartIcon = useSetRecoilState(cartClick)
-
-  // UseEffect to store into the local storage our Cart
-  useEffect(() => {
-    if (showCart?.length > 0) {
-      localStorage.removeItem('myCart')
-      useStoreCartToLocalStorage(showCart)
-    }
-  }, [showCart])
-
-  // If there is already a Cart in the local storage, then store it in recoile state
-  useEffect(() => {
-    const getCart = localStorage.getItem('myCart')
-    if (getCart) {
-      const cleanCart = JSON.parse(getCart)
-      const savedCart = cleanCart[cleanCart?.length - 1]
-      setShowCart(savedCart)
-    }
-  }, [])
-
-  const sumAllArticles = (cart) => {
-    let x = 0
-    cart.map((i, index) => {
-      x += i.qty
-    })
-    return x
-  }
 
   return (
     <ul
@@ -170,46 +110,6 @@ const Navigation = ({ isMenuOpen, setIsMenuOpen }) => {
           </li>
         )
       })}
-      <li className="container__header-nav-selectors" ref={selectorsNavigation}>
-        {shouldShowPersonasAtom && (
-          <div>
-            <Selectors props={personaConfig} />
-          </div>
-        )}
-        {shouldShowSegmentsAtom && (
-          <div>
-            <Selectors props={segmentConfig} />
-          </div>
-        )}
-        {/* Display the language select component */}
-        {shouldShowLanguageSelected && (
-          <div>
-            <Selectors props={languagesConfig} />
-          </div>
-        )}
-      </li>
-      {shouldShowCartIcon && (
-        <li
-          className={
-            cartOpenValue ? 'picto-cart picto-cart__active' : 'picto-cart'
-          }
-          onClick={() => {
-            setCartOpenValue(!cartOpenValue)
-            {
-              mobile && setIsMenuOpen(false)
-            }
-          }}
-          ref={cartIcon}
-        >
-          {!isDesktop ? <p>Cart</p> : <CartPicto />}
-          {/* Picto notification up the cart icon */}
-          {showCart?.length !== 0 && (
-            <div className="notification-cart">
-              <span>{sumAllArticles(showCart)}</span>
-            </div>
-          )}
-        </li>
-      )}
     </ul>
   )
 }
