+++ src/scss/partials/_globals.scss
@@ -23,6 +23,7 @@ body {
   overflow-y: auto;
   background-color: $background-color;
   font-family: $main-font;
+  max-width: 1600px;
 }
 
 h1,
