+++ src/scss/partials/_variables.scss
@@ -2,6 +2,7 @@
 // Colors
 $main-color: rgb(0, 0, 0);
 $secondary-color: rgb(0,61,255);
+$tertiary-color: rgb(5, 13, 35);
 
 // Background Color:
 $background-color: rgb(249, 249, 249);
